Python-based query building
