+++ backend/database/__init__.py
@@ -11,6 +11,8 @@ from database._row import (
     require_row,
     row_bool,
     row_int,
+    row_int_list_or_none,
+    row_int_or_none,
     row_str,
     row_str_list,
     row_str_or_none,
